requirements.txt
absl-py==2.1.0
alpaca-py==0.9.0
google-api-core==2.19.1
google-auth==2.23.4
google-cloud-core==2.4.1
google-cloud-pubsub==2.20.2
google-cloud-secret-manager==2.18.0
google-cloud-storage==2.10.0
googleapis-common-protos==1.63.0
grpcio==1.62.3
numpy==1.26.4
pandas==2.2.2
protobuf==4.25.3
pyyaml==6.0
requests==2.31.0
scikit-learn==1.4.2
tensorboard==2.16.2
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
transformers==4.40.0
joblib==1.4.2
scipy==1.13.1
threadpoolctl==3.5.0
typing-extensions==4.10.0
